database was changed

+++ database.py
@@ -8,7 +8,7 @@ class Database:
         self.connect_to_db()
 
     @staticmethod
-    def dataBaseMethod(func):
+    def data_base_method(func):
         def inner(self, *args, **kwargs):
             try:
                 res = func(self, *args, **kwargs)
@@ -30,103 +30,109 @@ class Database:
                                      port=port)
         self.cursor = self.conn.cursor()
 
-    def getCursor(self):
-        return self.cursor
-
-    @dataBaseMethod
-    def getValue(self, table, columns, condition=""):
+    @data_base_method
+    def get_value(self, table, columns, condition=""):
         sql = f"SELECT {columns} FROM {table}" + " " + condition
         self.cursor.execute(sql)
         return self.cursor.fetchall()
 
-    @dataBaseMethod
-    def updateValue(self, table, row, value, condition=""):
+    @data_base_method
+    def update_value(self, table, row, value, condition=""):
         sql = f"UPDATE {table} SET {row} = '{value}'" + " " + condition
         self.cursor.execute(sql)
         return 0
 
-    @dataBaseMethod
-    def deleteValue(self, table, condition=""):
+    @data_base_method
+    def delete_value(self, table, condition=""):
         sql = f"DELETE FROM {table}" + " " + condition
         self.cursor.execute(sql)
         return 0
 
-    @dataBaseMethod
-    def insertCart(self, customer_id, product_id, taken_value):
-        sql = f"INSERT INTO cart(customer_id, product_id, taken_value)  VALUES ({customer_id}, {product_id}, {taken_value})"
+    @data_base_method
+    def insert_cart(self, customer_id, product_id, taken_value):
+        sql = f"INSERT INTO cart(customer_id, product_id, taken_value) VALUES ({self.get_user_id(customer_id)}, {product_id}, {taken_value})"
         self.cursor.execute(sql)
         return 0
 
-    @dataBaseMethod
-    def insertUser(self, user_id, phone, area, deleted):
-        sql = f"INSERT INTO customer VALUES ({user_id}, '{phone}', '{area}', {deleted})"
+    @data_base_method
+    def insert_user(self, user_id, phone, area, deleted, mail):
+        sql = f"INSERT INTO customer(tg_id, phone_number, area, deleted, mail) VALUES ({user_id}, '{phone}', '{area}', {deleted}, {mail})"
         self.cursor.execute(sql)
         return 0
 
-    @dataBaseMethod
+    @data_base_method
     def free_request_no_fetch(self, req):
         self.cursor.execute(req)
 
-    @dataBaseMethod
+    @data_base_method
     def free_request_fetch(self, req):
         self.cursor.execute(req)
         return self.cursor.fetchall()
 
+    def get_user_id(self, user_id):
+        get_customer_tg_id = f"SELECT id FROM customer WHERE tg_id = {user_id}"
+        self.cursor.execute(get_customer_tg_id)
+        data = self.cursor.fetchone()
+        return data[0] if data is not None else None
+
     def change_phone(self, new_phone, user_id):
-        self.updateValue('customer', 'phone_number', new_phone, condition=f"WHERE tg_id = {user_id}")
+        self.update_value('customer', 'phone_number', new_phone, condition=f"WHERE id = {self.get_user_id(user_id)}")
 
     def delete_account(self, user_id):
-        self.updateValue('customer', 'deleted', 1, condition=f"WHERE tg_id = {user_id}")
+        self.update_value('customer', 'deleted', 1, condition=f"WHERE id = {self.get_user_id(user_id)}")
 
     def get_customer_data(self, user_id):
-        return self.getValue("customer", "*", condition=f"WHERE tg_id = {user_id}")
+        id = self.get_user_id(user_id)
+        return self.get_value("customer", "*", condition=f"WHERE id = {id if id is not None else -1}")
 
     def get_product_data(self):
-        return self.getValue("product", "*")
+        return self.get_value("product", "*")
 
     def get_product_data_by_product_id(self, product_id):
-        return self.getValue("product", "*", condition=f"WHERE product_id = {product_id}")
+        return self.get_value("product", "*", condition=f"WHERE product_id = {product_id}")
 
     def get_remainder_product(self, product_id):
-        return self.getValue("product", "remainder", condition=f"WHERE product_id = {product_id}")
+        return self.get_value("product", "remainder", condition=f"WHERE product_id = {product_id}")
 
     def get_product_in_cart(self, user_id, product_id):
-        return self.getValue("cart", "*", condition=f"WHERE customer_id = {user_id} AND product_id = {product_id}")
+        return self.get_value("cart", "*",
+                              condition=f"WHERE customer_id = {self.get_user_id(user_id)} AND product_id = {product_id}")
 
     def change_product_in_cart_taken_value(self, taken_value, new_value, user_id, product_id):
-        self.updateValue("cart", "taken_value", taken_value + new_value,
-                         condition=f"WHERE customer_id = {user_id} AND product_id = {product_id}")
+        self.update_value("cart", "taken_value", taken_value + new_value,
+                          condition=f"WHERE customer_id = {user_id} AND product_id = {product_id}")
 
     def get_cart_data(self, user_id):
         return self.free_request_fetch("SELECT product.name, product.cost, cart.taken_value "
                                        "FROM cart INNER JOIN product ON product.product_id = cart.product_id "
-                                       f"WHERE customer_id = {user_id}")
+                                       f"WHERE customer_id = {self.get_user_id(user_id)}")
 
     def get_product_data_by_user_id(self, user_id):
         return self.free_request_fetch("SELECT product.name, cart.product_id "
                                        "FROM cart INNER JOIN product ON cart.product_id = product.product_id "
-                                       f"WHERE cart.customer_id = {user_id}")
+                                       f"WHERE cart.customer_id = {self.get_user_id(user_id)}")
 
     def delete_from_cart(self, product_id, user_id):
         self.free_request_no_fetch(f"DELETE FROM cart "
-                                   f"WHERE cart.product_id = {product_id} AND cart.customer_id = {user_id}")
+                                   f"WHERE cart.product_id = {product_id} AND cart.customer_id = {self.get_user_id(user_id)}")
 
     def get_product_name_taken_value_remainder(self, user_id):
         return self.free_request_fetch(f"SELECT product.name, cart.taken_value, product.remainder "
                                        f"FROM cart INNER JOIN product ON cart.product_id = product.product_id "
-                                       f"WHERE cart.customer_id = {user_id}")
+                                       f"WHERE cart.customer_id = {self.get_user_id(user_id)}")
 
     def insert_order(self, user_id, date, payment_method):
         return self.free_request_fetch(f"INSERT INTO orders(customer_id, date, payment_method, status) "
-                                       f"VALUES({user_id}, '{date}', '{payment_method}', 0) RETURNING id;")
+                                       f"VALUES({self.get_user_id(user_id)}, '{date}', '{payment_method}', 0) RETURNING id;")
 
     def get_all_cart_data(self, user_id):
-        return self.free_request_fetch(f"SELECT product_id, taken_value FROM cart WHERE customer_id = {user_id}")
+        return self.free_request_fetch(
+            f"SELECT product_id, taken_value FROM cart WHERE customer_id = {self.get_user_id(user_id)}")
 
     def insert_position(self, user_id, request):
         self.free_request_no_fetch(f"INSERT INTO positions(order_id, product_id, taken_value) VALUES {request}")
 
-        self.free_request_no_fetch(f"DELETE FROM cart WHERE customer_id = {user_id}")
+        self.free_request_no_fetch(f"DELETE FROM cart WHERE customer_id = {self.get_user_id(user_id)}")
 
     def update_remainders_products(self, cart_data):
         for row in cart_data:
@@ -139,7 +145,10 @@ class Database:
             "orders.date, orders.payment_method, orders.status "
             "FROM orders INNER JOIN positions ON orders.id = positions.order_id "
             "INNER JOIN product ON positions.product_id = product.product_id "
-            f"WHERE orders.customer_id = {user_id}")
+            f"WHERE orders.customer_id = {self.get_user_id(user_id)}")
 
     def get_product_ids(self):
-        return [str(row[0]) for row in self.getValue("product", "product_id")]
+        return [str(row[0]) for row in self.get_value("product", "product_id")]
+
+    def change_area(self, message, area):
+        self.update_value('customer', 'area', area, condition=f"WHERE id = {self.get_user_id(message.chat.id)}")
